requirements.txt
argparse
asyncio
datetime
jwt
schedule
envoy_utils
httpx
bs4
urllib3
apscheduler
flask
requests
influxdb-client
psycopg2-binary